refactor(#32) :: 폴더 위치 register 하위로 이동
