+++ src/pages/register/surveyEnd/index.tsx
