+++ src/pages/register/surveyEnd/style.ts
