+++ src/pages/register/surveyFood/style.ts
