+++ src/pages/register/surveyMBTI/index.tsx
