+++ src/pages/register/surveyMBTI/style.ts
