+++ src/pages/register/surveyStyle/index.tsx
