+++ src/pages/register/surveyStyle/style.ts
