+++ src/routes/index.tsx
@@ -4,10 +4,10 @@ import Splash from '../pages/splash/';
 import Auth from 'pages/auth';
 import Terms from 'pages/terms';
 import Info from 'pages/register/info';
-import SurveyEnd from '../pages/surveyEnd';
-import SurveyStyle from 'pages/surveyStyle';
-import SurveyFood from 'pages/surveyFood';
-import SurveyMBTI from 'pages/surveyMBTI';
+import SurveyEnd from '../pages/register/surveyEnd';
+import SurveyStyle from 'pages/register/surveyStyle';
+import SurveyFood from 'pages/register/surveyFood';
+import SurveyMBTI from 'pages/register/surveyMBTI';
 
 export default function Router() {
   return useRoutes([
